configured jackson to always autobox numbers as long — switch ids to long

+++ backend/src/main/java/com/smartmerge/controller/PullRequestController.java
@@ -19,7 +19,7 @@ public class PullRequestController {
     private final PullRequestService pullRequestService;
 
     @GetMapping("/repository/{repoId}")
-    public ResponseEntity<List<PullRequest>> getPullRequestsByRepoId(@PathVariable int repoId) {
+    public ResponseEntity<List<PullRequest>> getPullRequestsByRepoId(@PathVariable long repoId) {
         return ResponseEntity.ok(pullRequestService.getPullRequestsByRepoId(repoId));
     }
 }
