+++ backend/src/main/java/com/smartmerge/repository/InstallationRepository.java
@@ -3,6 +3,6 @@ package com.smartmerge.repository;
 import com.smartmerge.model.Installation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstallationRepository extends JpaRepository<Installation, Integer> {
+public interface InstallationRepository extends JpaRepository<Installation, Long> {
 
 }
