+++ backend/src/main/java/com/smartmerge/repository/ProfileRepository.java
@@ -3,6 +3,6 @@ package com.smartmerge.repository;
 import com.smartmerge.model.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
 
 }
