+++ backend/src/main/java/com/smartmerge/repository/PullRequestRepository.java
@@ -2,17 +2,15 @@ package com.smartmerge.repository;
 
 import com.smartmerge.model.PullRequest;
 import jakarta.transaction.Transactional;
-
 import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PullRequestRepository extends JpaRepository<PullRequest, Long> {
 
-    List<PullRequest> findAllByRepoId(int repoId);
+    List<PullRequest> findAllByRepoId(long repoId);
 
     @Transactional
-    void deleteAllByInstallationId(int id);
+    void deleteAllByInstallationId(long id);
 
-    void deleteAllByRepoIdIn(List<Integer> repoIds);
+    void deleteAllByRepoIdIn(List<Long> repoIds);
 }
