+++ backend/src/main/java/com/smartmerge/service/RepoService.java
@@ -22,13 +22,13 @@ public class RepoService {
         return repoRepository.saveAll(repos);
     }
 
-    public List<Repo> getReposByUserId(int userId) {
+    public List<Repo> getReposByUserId(long userId) {
         return repoRepository.findAllByUserId(userId)
                 .orElseThrow(() -> new NoSuchElementException("User Id not found"));
     }
 
     @Transactional
-    public void deleteAllRepos(List<Integer> repoIds) {
+    public void deleteAllRepos(List<Long> repoIds) {
         // delete all PRs with any of these repoIds
         pullRequestRepository.deleteAllByRepoIdIn(repoIds);
         repoRepository.deleteAllById(repoIds);
